refactor: BoxOfficeDataLoader, MovieInfoDataLoader, MoviePosterImageLoader 3가지 타입으로 기능 분리
